[iso] image extraction at last, for both UDF and ISO9660
* removed the need for ftruncate and CEILING
* revert "bad offset computation ICB" 'fix' (there never was a problem)
* use 64bit types and calls where needed
* minor utf8 fixes

+++ src/libcdio/udf/udf_private.h
@@ -33,7 +33,7 @@
 struct udf_s {
   bool          b_stream;         /* Use stream pointer, else use 
 				    p_cdio.  */
-  ssize_t               i_position; /* Position in file if positive. */
+  off64_t               i_position; /* Position in file if positive. */
   CdioDataSource_t      *stream;  /* Stream pointer if stream */
   CdIo_t                *cdio;    /* Cdio pointer if read device */
   anchor_vol_desc_ptr_t anchor_vol_desc_ptr;
